Compile fix

+++ renderdoc/driver/vulkan/wrappers/vk_resource_funcs.cpp
@@ -1104,7 +1104,7 @@ bool WrappedVulkan::Serialise_vkFlushMappedMemoryRanges(SerialiserType &ser, VkD
         size_t size = size_t(RDCMIN(it->finish(), finish) - offs);
 
         // cap the size by the mapped memory region
-        size = RDCMIN(MemRange.size, size);
+        size = RDCMIN((size_t)MemRange.size, size);
 
         memcpy(MappedData + offs, m_MaskedMapData.data() + offs, size);
       }
